pyproject.toml
[project]
name = "news_headlines_bot"
version = "0.1.0"
description = "Telegram bot for news generation"
authors = [{name = "Pavel"}]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "aiogram>=3.7",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.18",
    "beautifulsoup4>=4.14",
    "httpx>=0.27",
    "apscheduler>=3.11",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "psycopg2-binary>=2.9",
    "rss-parser>=2.1.1",
    "feedparser>=6.0.12",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = ["watchdog>=6.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Если Alembic пытается читать этот файл
[tool.alembic]
alembic_ini_path = "alembic.ini"
version_path = "migrations/versions"